Add Il2Cpp support:
- Conditionally add necessary changes for Il2Cpp.
- Add the EnlargeCategories() function to re-implement Array.Resize() in
an Il2Cpp-friendly manner. There's probably a way to call Array.Resize()
from Il2Cpp code but this was easier than figuring it out.
- In RetentionAmountSlider, re-implement the functionality of
base.OnValueChanged() in OnValueChanged(). Similar to the above, this
was easier than understanding why the call wasn't working.
- Don't worry about adding a listener to modSettingsButton; this is
handled for us in SettingsScreen.Start(), which will run later.